# app/Http/Resources/Admin/ServicePackageResource.php
<?php

namespace App\Http\Resources\Admin;

use Illuminate\Http\Resources\Json\JsonResource;

class ServicePackageResource extends JsonResource
{
    public function toArray($request)
    {
        return parent::toArray($request);
    }
}