# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Resolvy API Documentation",
        "description": "API documentation for the Resolvy helpdesk application",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/v1/service-packages": {
            "get": {
                "tags": [
                    "Service Packages"
                ],
                "summary": "Get list of service packages",
                "description": "Returns a paginated list of all service packages",
                "operationId": "getServicePackagesList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ServicePackageResource"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service packages retrieved successfully."
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Service Packages"
                ],
                "summary": "Create a new service package",
                "description": "Creates a new service package record",
                "operationId": "storeServicePackage",
                "requestBody": {
                    "description": "Service package data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreServicePackageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service package created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ServicePackageResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service package created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/service-packages/{id}": {
            "get": {
                "tags": [
                    "Service Packages"
                ],
                "summary": "Get service package information",
                "description": "Returns service package data",
                "operationId": "getServicePackageById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of service package to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ServicePackageResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service package retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Service Packages"
                ],
                "summary": "Update an existing service package",
                "description": "Updates a service package record",
                "operationId": "updateServicePackage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of service package to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Service package data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateServicePackageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Service package updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ServicePackageResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service package updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Service Packages"
                ],
                "summary": "Delete an existing service package",
                "description": "Deletes a service package record",
                "operationId": "deleteServicePackage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of service package to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service package deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "example": null
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service package deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginationMeta": {
                "title": "Pagination Meta",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "path": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/customers"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 150
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "title": "Pagination Links",
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/customers?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/customers?page=10"
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/customers?page=2"
                    }
                },
                "type": "object"
            },
...
            "StoreServicePackageRequest": {
                "title": "Store Service Package Request",
                "required": [
                    "name",
                    "price"
                ],
                "properties": {
                    "name": {
                        "description": "The name of the service package",
                        "type": "string",
                        "example": "Gold Tier Support"
                    },
                    "description": {
                        "description": "A description of the package",
                        "type": "string",
                        "example": "24/7 phone and email support"
                    },
                    "price": {
                        "description": "The price of the package",
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    }
                },
                "type": "object"
            },
            "UpdateServicePackageRequest": {
                "title": "Update Service Package Request",
                "properties": {
                    "name": {
                        "description": "The name of the service package",
                        "type": "string",
                        "example": "Platinum Tier Support"
                    },
                    "description": {
                        "description": "A description of the package",
                        "type": "string",
                        "example": "Dedicated support agent and 24/7 phone support"
                    },
                    "price": {
                        "description": "The price of the package",
                        "type": "number",
                        "format": "float",
                        "example": 199.99
                    }
                },
                "type": "object"
            },
            "ServicePackageResource": {
                "title": "Service Package Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Gold Tier Support"
                    },
                    "description": {
                        "type": "string",
                        "example": "24/7 phone and email support"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}